+++ Assets/Scripts/HP/IDamageable.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 public interface IDamageable 
 {
     public float Hp { get; set; }
-    void TakeDamage(float damage, LayerMask mask);
+    void TakeDamage(int damage, LayerMask mask);
     void CheckDeath();
 }
